Update ddqn and simple_pg files
